GP_requirements.txt
aicsimageio==4.14.0
aicspylibczi==3.2.1
aiobotocore==2.5.4
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aioitertools==0.12.0
aiosignal==1.3.2
alabaster==1.0.0
annotated-types==0.7.0
anyio==4.8.0
app-model==0.3.1
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asciitree==0.3.3
asttokens==3.0.0
async-lru==2.0.4
async-timeout==5.0.1
attrs==25.1.0
babel==2.17.0
beautifulsoup4==4.13.3
bleach==6.2.0
botocore==1.31.17
build==1.2.2.post1
cachey==0.2.1
cellpose==3.1.1.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cupy-cuda12x==13.3.0
cycler==0.12.1
czifile==2019.7.2
dask==2025.2.0
debugpy==1.8.12
decorator==5.1.1
defusedxml==0.7.1
distributed==2025.2.0
docstring_parser==0.16
docutils==0.21.2
elementpath==4.7.0
exceptiongroup==1.2.2
executing==2.2.0
fasteners==0.19
fastjsonschema==2.21.1
fastremap==1.15.0
fastrlock==0.8.3
filelock==3.13.1
flexcache==0.3
flexparser==0.4
fonttools==4.56.0
fqdn==1.5.1
freetype-py==2.5.1
frozenlist==1.5.0
fsspec==2023.6.0
h11==0.14.0
HeapDict==1.0.1
hsluv==5.0.4
httpcore==1.0.7
httpx==0.28.1
idna==3.10
imagecodecs==2024.12.30
imageio==2.37.0
imageio-ffmpeg==0.6.0
imagesize==1.4.1
importlib_metadata==8.6.1
in-n-out==0.2.1
ipykernel==6.29.5
ipython==8.32.0
ipywidgets==8.1.5
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.5
jmespath==1.0.1
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
jupyterthemes==0.20.0
kiwisolver==1.4.8
lazy_loader==0.4
lesscpy==0.15.1
llvmlite==0.44.0
locket==1.0.0
lxml==4.9.4
magicgui==0.10.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
matplotlib-scalebar==0.8.1
mdurl==0.1.2
meshio==5.3.5
mistune==3.1.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
napari==0.5.6
napari-animation==0.0.8
napari-console==0.1.3
napari-czifile2==0.2.7
napari-matplotlib==3.0.0
napari-meshio==0.0.1
napari-plugin-engine==0.2.0
napari-plugin-manager==0.1.4
napari-svg==0.2.1
natsort==8.4.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nd2reader==3.3.1
nest-asyncio==1.6.0
networkx==3.4.2
notebook==7.3.2
notebook_shim==0.2.4
npe2==0.7.7
numba==0.61.0
numcodecs==0.13.1
numpy==1.26.4
numpydoc==1.8.0
ome-types==0.5.3
ome-zarr==0.10.3
opencv-python-headless==3.4.18.65
overrides==7.7.0
packaging==24.2
palettable==3.3.3
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
partd==1.4.2
PartSegCore_compiled_backend==0.15.11
pillow==11.1.0
PIMS==0.7
Pint==0.24.4
platformdirs==4.3.6
ply==3.11
pooch==1.8.2
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.2.1
psutil==7.0.0
psygnal==0.12.0
pure_eval==0.2.3
pyconify==0.2.1
pycparser==2.22
pydantic==2.10.6
pydantic-compat==0.1.2
pydantic_core==2.27.2
Pygments==2.19.1
PyOpenGL==3.1.9
pyparsing==3.2.1
pyproject_hooks==1.2.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
python-dateutil==2.9.0.post0
python-json-logger==3.2.1
pytz==2025.1
pywin32==308
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==26.2.1
qtconsole==5.6.1
QtPy==2.4.3
readlif==0.6.5
referencing==0.36.2
requests==2.32.3
resource-backed-dask-array==0.1.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
roifile==2024.9.15
rpds-py==0.22.3
s3fs==2023.6.0
scikit-image==0.25.1
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
seaborn-image==0.10.0
Send2Trash==1.8.3
shellingham==1.5.4
six==1.17.0
slicerator==1.1.0
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.6
Sphinx==8.1.3
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
stack-data==0.6.3
superqt==0.7.1
sympy==1.13.1
tabulate==0.9.0
tblib==3.0.0
terminado==0.18.1
threadpoolctl==3.5.0
tifffile==2023.2.28
tinycss2==1.4.0
tomli==2.2.1
tomli_w==1.2.0
toolz==1.0.0
torch==2.6.0+cu124
torchaudio==2.6.0+cu124
torchvision==0.21.0+cu124
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
triangle==20250106
typer==0.15.1
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2025.1
uri-template==1.3.0
urllib3==1.26.20
vispy==0.14.3
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.13
wrapt==1.17.2
xarray==2025.1.2
XlsxWriter==3.2.2
xmlschema==3.4.3
xmltodict==0.14.2
xsdata==24.3.1
yarl==1.18.3
zarr==2.15.0
zict==3.0.0
zipp==3.21.0
